Добавил сохранение в БД списка. Сохраняем каждое изменение. попал в список - insert, переместил в корзину - update, удалил - delete. Следующий шаг - загрузка списка из бд

+++ app/src/main/java/ru/alexandertsebenko/shoplist2/datamodel/ShopList.java
@@ -14,7 +14,7 @@ public class ShopList {
     private long dateMilis;
     List<ProductInstance> prodList = new ArrayList<>();
 
-    public ShopList(int id, long date, String name){
+    public ShopList(long id, long date, String name){
         this.id = id;
         this.dateMilis = date;
         this.name = name;
